fix #40

+++ app/Http/Controllers/AdminContoller.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Mail\DeleteAccountMail;
+use App\Models\Session;
 use App\Models\User;
 use App\Validators\ValidatesAdminRequests;
 use Illuminate\Http\JsonResponse;
@@ -10,7 +11,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Mail;
 
-class AccountsController extends Controller
+class AdminController extends Controller
 {
     use ValidatesAdminRequests;
 
@@ -33,12 +34,17 @@ class AccountsController extends Controller
     /**
      * Show user.
      *
-     * @param int $id
+     * @param Request $request
+     * @param int     $id
      *
      * @return JsonResponse
      */
-    public function show($id)
+    public function show(Request $request, $id)
     {
+        if ($this->checkID($request, $id)) {
+            return $this->checkID($request, $id);
+        }
+
         $user = User::findOrFail($id);
 
         return $this->respondSuccess(
@@ -59,6 +65,10 @@ class AccountsController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if ($this->checkID($request, $id)) {
+            return $this->checkID($request, $id);
+        }
+
         $user = User::findOrFail($id);
 
         $this->validateUpdate($request, $user);
@@ -84,15 +94,22 @@ class AccountsController extends Controller
      * Delete user
      * Revokes all sessions.
      *
-     * @param int $id
+     * @param Request $request
+     * @param int     $id
      *
      * @return JsonResponse
      */
-    public function delete($id)
+    public function delete(Request $request, $id)
     {
+        if ($this->checkID($request, $id)) {
+            return $this->checkID($request, $id);
+        }
+
         $user = User::findOrFail($id);
         $user->sessions()->delete();
+
         Mail::to($user->email)->send(new DeleteAccountMail($user));
+
         $user->delete();
 
         return $this->respondSuccess(
@@ -104,12 +121,17 @@ class AccountsController extends Controller
     /**
      * Show products.
      *
-     * @param int $id
+     * @param Request $request
+     * @param int     $id
      *
      * @return JsonResponse
      */
-    public function showProducts($id)
+    public function showProducts(Request $request, $id)
     {
+        if ($this->checkID($request, $id)) {
+            return $this->checkID($request, $id);
+        }
+
         $products = User::findOrFail($id)->products();
 
         return $this->respondSuccess(
@@ -122,12 +144,17 @@ class AccountsController extends Controller
     /**
      * Show sessions.
      *
-     * @param int $id
+     * @param Request $request
+     * @param int     $id
      *
      * @return JsonResponse
      */
-    public function showSessions($id)
+    public function showSessions(Request $request, $id)
     {
+        if ($this->checkID($request, $id)) {
+            return $this->checkID($request, $id);
+        }
+
         $sessions = User::findOrFail($id)->sessions();
 
         return $this->respondSuccess(
@@ -141,18 +168,49 @@ class AccountsController extends Controller
      * Revoke an session.
      *
      * @param Request $request
+     * @param int     $id
      *
      * @return JsonResponse
      */
-    public function revokeSession(Request $request)
+    public function revokeSession(Request $request, $id)
     {
         $this->validateRevoke($request);
+        if ($this->checkID($request, $id)) {
+            return $this->checkID($request, $id);
+        }
 
-        Session::findOrFail($request->get('session_uuid'))->delete();
+        $session = Session::findOrFail($request->get('session_uuid'))->whereUserId($id)->first();
+
+        if (!$session) {
+            return $this->respondError(
+                'model not found',
+                'CLIENT_ERROR_NOT_FOUND'
+            );
+        }
+
+        $session->delete();
 
         return $this->respondSuccess(
             'user session revoked',
             'SUCCESS_OK'
         );
+    }
+
+    /**
+     * Validate target isn't current account.
+     *
+     * @param Request $request
+     * @param string  $id_request
+     *
+     * @return null|JsonResponse
+     */
+    protected function checkID(Request $request, $id_request)
+    {
+        if ($id_request == $request->user_id) {
+            return $this->respondError(
+                'can\'t perform actions on your own account',
+                'CLIENT_ERROR_BAD_REQUEST'
+            );
+        }
     }
 }
